gst: don't use dash as API version separator in plugin filename
Since d59e26d37dd6be535460d6831f3d4ad6666808b5 gstreamer commit, dash
are allowed in plugin filename, which means the first number of API
version was used for the building of the name of the _get_desc plugin
function. The loading of the plugin failed because it did not find
_get_desc.

This commit also fixes the plugin name to be aravis, aravisrsrc being
the source element.