Ignore duplicate packets in ArvGvStream (#313)
With this change, ArvGvStream will discard duplicate packets
for packet_ids already received. The previous behaviour was
to use the payload from the duplicate packet, replacing the
previously received data.

Discarding duplicate packets acts as a workaround for an
issue with PtGrey BlackFly cameras which sometimes send
duplicate packets with incorrect data.

See issue #311

It doesn't fix what seems a second class of wrong packets, which are late resent packets with not only the wrong packet id, but also the wrong frame id. The resulting artefacts are wrong lines at the bottom of the frame.