gst: ignore region setting if it already has the right values
On some devices, Width and Height are read only registers. Trying to set them
returns an error. As an error is fatal in the gstreamer element, we have to
not call set_region if the region is already set to the right values, which is
what happens when the registers are read only.